Move `BreedsPresenter` to its own file
