+++ CatsFoundation/CatsFoundationTests/Breeds Presentation/BreedsPresenterTests.swift
@@ -5,22 +5,6 @@
 @testable import CatsFoundation
 import XCTest
 
-struct BreedsPresenter {
-    let title = "Breeds"
-
-    func map(_ breed: Breed) -> PresentableBreed {
-        PresentableBreed(
-            breedID: "id: \(breed.id)",
-            name: breed.name,
-            description: breed.description,
-            temperament: breed.temperament,
-            origin: breed.origin,
-            lifeSpan: "\(breed.lifeSpan) average life span",
-            imageURL: breed.image
-        )
-    }
-}
-
 final class BreedsPresenterTests: XCTestCase {
     func test_hasTitle() {
         let sut = BreedsPresenter()
